backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
chromadb==0.4.18
sentence-transformers==2.2.2
requests==2.31.0
beautifulsoup4==4.12.2
anthropic==0.20.0
openai==1.6.1
python-dotenv==1.0.0
Pillow==10.1.0
lxml==4.9.3
